requirements.txt
appdirs==1.4.3
cycler==0.10.0
Django==1.10.6
djangorestframework==3.6.3
h5py==2.7.0
Keras==2.0.5
matplotlib==2.0.0
netifaces==0.10.5
numpy==1.13.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
protobuf==3.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
sklearn==0.0
tensorflow==1.0.1
Theano==0.9.0
